requirements.txt
pandas==1.0.1
streamlit==0.59.0
Pillow==7.1.2
streamlit==0.59.0
